Updated FJ device - symbol and package
Symbol pins are now correctly named A0 and A1, they used to have A0/TX0
and A1/RX0, which is incorrect and left over from the mega pro designs.

Also, with this update to the device came a small update to the package.
Only thing effected was the drill hits for the brackets are now the
correct "tighter" size of 0.11".